DateTime: refactor directory structures
